Cleanup ui package structure
