+++ app/src/main/kotlin/com/ayatk/biblio/ui/BaseActivity.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2016-2017 Aya Tokikaze. All Rights Reserved.
  */
 
-package com.ayatk.biblio.ui.activity
+package com.ayatk.biblio.ui
 
 import android.os.Bundle
 import android.support.annotation.IdRes
