+++ app/src/main/kotlin/com/ayatk/biblio/ui/ViewModel.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2016-2017 Aya Tokikaze. All Rights Reserved.
  */
 
-package com.ayatk.biblio.viewmodel
+package com.ayatk.biblio.ui
 
 interface ViewModel {
