+++ app/src/main/kotlin/com/ayatk/biblio/ui/detail/info/NovelInfoFragment.kt
@@ -1,8 +1,8 @@
 /*
- * Copyright (c) 2016-2017 Aya Tokikaze. All Rights Reserved.
+ * Copyright (c) 2016-2017. Aya Tokikaze. All Rights Reserved.
  */
 
-package com.ayatk.biblio.ui.fragment
+package com.ayatk.biblio.ui.detail.info
 
 import android.content.Context
 import android.os.Bundle
@@ -12,8 +12,8 @@ import android.view.ViewGroup
 import com.ayatk.biblio.databinding.FragmentNovelInfoBinding
 import com.ayatk.biblio.model.Novel
 import com.ayatk.biblio.repository.library.LibraryRepository
+import com.ayatk.biblio.ui.BaseFragment
 import com.ayatk.biblio.ui.util.helper.Navigator
-import com.ayatk.biblio.viewmodel.NovelInfoViewModel
 import org.parceler.Parcels
 import javax.inject.Inject
 
@@ -28,7 +28,8 @@ class NovelInfoFragment : BaseFragment() {
   lateinit var libraryRepository: LibraryRepository
 
   private val novel: Novel by lazy {
-    Parcels.unwrap<Novel>(arguments?.getParcelable(BUNDLE_ARGS_NOVEL))
+    Parcels.unwrap<Novel>(arguments?.getParcelable(
+        BUNDLE_ARGS_NOVEL))
   }
 
   private lateinit var viewModel: NovelInfoViewModel
@@ -36,7 +37,8 @@ class NovelInfoFragment : BaseFragment() {
   override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                             savedInstanceState: Bundle?): View? {
     binding = FragmentNovelInfoBinding.inflate(inflater, container, false)
-    viewModel = NovelInfoViewModel(navigator, libraryRepository, novel)
+    viewModel = NovelInfoViewModel(navigator, libraryRepository,
+        novel)
     binding.viewModel = viewModel
 
     return binding.root
@@ -58,7 +60,8 @@ class NovelInfoFragment : BaseFragment() {
     fun newInstance(novel: Novel): NovelInfoFragment {
       return NovelInfoFragment().apply {
         arguments = Bundle().apply {
-          putParcelable(BUNDLE_ARGS_NOVEL, Parcels.wrap(novel))
+          putParcelable(
+              BUNDLE_ARGS_NOVEL, Parcels.wrap(novel))
         }
       }
     }
