--- app/src/main/kotlin/com/ayatk/biblio/ui/fragment/SearchFragment.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2016-2017 Aya Tokikaze. All Rights Reserved.
- */
-
-package com.ayatk.biblio.ui.fragment
-
-import android.content.Context
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.ayatk.biblio.databinding.FragmentSearchBinding
-
-class SearchFragment : BaseFragment() {
-
-  lateinit var binding: FragmentSearchBinding
-
-  override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                            savedInstanceState: Bundle?): View? {
-    binding = FragmentSearchBinding.inflate(inflater, container, false)
-    return binding.root
-  }
-
-  override fun onAttach(context: Context?) {
-    super.onAttach(context)
-    component().inject(this)
-  }
-
-  companion object {
-    fun newInstance(): SearchFragment = SearchFragment()
-  }
-}
