+++ app/src/main/kotlin/com/ayatk/biblio/ui/search/SearchResultItemViewModel.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2016-2017. Aya Tokikaze. All Rights Reserved.
  */
 
-package com.ayatk.biblio.viewmodel
+package com.ayatk.biblio.ui.search
 
 import android.databinding.BaseObservable
 import android.databinding.Bindable
@@ -12,6 +12,7 @@ import com.ayatk.biblio.model.Library
 import com.ayatk.biblio.model.Novel
 import com.ayatk.biblio.model.enums.NovelState
 import com.ayatk.biblio.repository.library.LibraryRepository
+import com.ayatk.biblio.ui.ViewModel
 import com.ayatk.biblio.util.FORMAT_yyyyMMdd_kkmm
 import io.reactivex.rxkotlin.subscribeBy
 import timber.log.Timber
