+++ app/src/main/res/layout/activity_novel_detail.xml
@@ -7,13 +7,13 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.activity.NovelDetailActivity">
+    tools:context=".ui.detail.NovelDetailActivity">
 
   <data>
 
     <variable
         name="viewModel"
-        type="com.ayatk.biblio.viewmodel.NovelDetailViewModel"/>
+        type="com.ayatk.biblio.ui.detail.NovelDetailViewModel"/>
   </data>
 
   <android.support.design.widget.CoordinatorLayout
