+++ app/src/main/res/layout/fragment_novel_info.xml
@@ -12,7 +12,7 @@
   <data>
     <variable
         name="viewModel"
-        type="com.ayatk.biblio.viewmodel.NovelInfoViewModel"/>
+        type="com.ayatk.biblio.ui.detail.info.NovelInfoViewModel"/>
   </data>
 
   <ScrollView
