+++ app/src/main/res/layout/view_library_item.xml
@@ -12,7 +12,7 @@
 
     <variable
         name="viewModel"
-        type="com.ayatk.biblio.viewmodel.LibraryItemViewModel"/>
+        type="com.ayatk.biblio.ui.home.library.LibraryItemViewModel"/>
   </data>
 
   <android.support.constraint.ConstraintLayout
