+++ app/src/main/res/layout/view_table_item.xml
@@ -14,7 +14,7 @@
 
     <variable
         name="viewModel"
-        type="com.ayatk.biblio.viewmodel.NovelTableItemViewModel"/>
+        type="com.ayatk.biblio.ui.detail.table.NovelTableItemViewModel"/>
   </data>
 
   <android.support.constraint.ConstraintLayout
